requirements.txt
Flask>=3.1,<4.0
Werkzeug>=3.1,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.1,<5.0
SQLAlchemy>=2.0,<3.0
alembic>=1.18,<2.0
python-dotenv>=1.0,<2.0
supabase>=2.28,<3.0
requests>=2.31,<3.0
openai>=2.32,<3.0